Delete: サーバーサイドからSkyHopper初期構築時のChefServer構築機能を削除
